refactor(build): extract enrichment-queue pure module from build-artifacts (#2042)
